+++ CHANGELOG.md
@@ -16,6 +16,18 @@ Types of changes
 
 # Latch SDK Changelog
 
+## 2.28.0 - 2023-07-25
+
+### Added
+
+- Tasks explicitly request ephemeral storage
+- `custom_task` and `custom_memory_optimized_task` allow selecting storage size
+- `custom_memory_optimized_task` functionality merged into `custom_task`
+
+### Deprecated
+
+- `custom_memory_optimized_task`
+
 ## 2.27.4 - 2023-07-18
 
 ### Changed